+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 # 1.0.1 (unreleased)
 - Make current personal data in read-mode after adding it.
 - Avoid to recreate food if OFF has already be stored.
+- Insensitive case search on foods.
+- Sort foods by last used.
 
 # 1.0.0
 - Programs, exercises, foods and personal data can be set.